At the Navbar added Private and Public components
