+++ src/pages/Layout.jsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import { Outlet } from 'react-router-dom'
-import Navbar from '../components/Navbar';
+import Navbar from '../components/UI/navbar/Navbar';
 
 function Layout(props) {
     return (
